Point to `benizi` forks by default
I think it's one of the stupidest decisions ever that package names in
Go include an absolute reference to the source repository. It makes it
stupidly complicated to use a custom fork (requires subbing out all
references within the fork and setting up custom directory structures
that shouldn't need to exist).